228 Add mdl_customcert_issues to compoundindexes
For mod_customcert rows with duplicate (userid, customcertid) are not
allowed although this is not specified in mdl_customcert_issues's
table definition.  When duplicates are present the message "Did you
remember to make the first column something unique in your call to
get_records? Duplicate value ... found in column 'id'." occurs.  This
commit adds handling for this compound index.